SourceCode/requirements.txt
diffusers==0.26.3
Flask==3.0.2
Flask_Cors==4.0.0
Pillow==10.2.0
torch==2.2.0
flask_socketio
transformers
accelerate
asgiref
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity